% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getBAMCoverage}
\alias{getBAMCoverage}
\title{Generate a coverage object from a BAM file}
\usage{
getBAMCoverage(bamfilename, regions)
}
\arguments{
\item{bamfilename}{path to the BAM file}
}
\value{
coverage object for whole BAM file
}
\description{
Generate a coverage object from a BAM file
}
\examples{
getBAMCoverage("/myproject/myBAM.bam")
}

