% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getBAMGRanges}
\alias{getBAMGRanges}
\title{Generate counts from a set of windows in a GRanges object}
\usage{
getBAMGRanges(bamfilename, peaksummits_gr, cores)
}
\arguments{
\item{bamfilename}{path to the BAM file}

\item{GRanges}{object of windows}
}
\value{
list of read counts for windows in each chromosome in order of the GRanges object
}
\description{
Generate counts from a set of windows in a GRanges object
}
\examples{
getBAMCoverage("/myproject/myBAM.bam",myGRanges)
}

