% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getBAMWindows}
\alias{getBAMWindows}
\title{Generate counts from a set of windows from a BAM file from a given set of chromosomes}
\usage{
getBAMWindows(bamfilename, regions, WinSz, cores)
}
\arguments{
\item{bamfilename}{path to the BAM file}

\item{regions}{a list of chromosomes (seqnames) to provide window counts for}

\item{cores}{number of cores to parallelize using}

\item{window}{size (window suze should be an odd integer value)}
}
\value{
list of read counts for windows in each chromosome
getBAMCoverage("/myproject/myBAM.bam",c("chr22",257))
}
\description{
Generate counts from a set of windows from a BAM file from a given set of chromosomes
}

