% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getMapabilityCoverage}
\alias{getMapabilityCoverage}
\title{Return coverage object for mapability file}
\usage{
getMapabilityCoverage(MapabilityBwFile)
}
\arguments{
\item{MapabilityBwFile}{path to the mapability BigWig File}
}
\value{
mapability coverage object filted for mapability of "1" regions
}
\description{
Return coverage object for mapability file
}
\examples{
getBAMCoverage("/myproject/myMapability.bw")
}

