% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getMapabilityGRanges}
\alias{getMapabilityGRanges}
\title{Return normalized mapability based on provided regions}
\usage{
getMapabilityGRanges(MapabilityBwFile, regions, ReadLen, cores)
}
\arguments{
\item{MapabilityBwFile}{path to the mapability BigWig File}

\item{cores}{number of cores to parallelize using}
}
\value{
mapability coverage object filtered for mapability of "1" regions
}
\description{
Return normalized mapability based on provided regions
}
\examples{
getBAMCoverage("/myproject/myMapability.bw",myGRanges,36)
}

