% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getMapabilityWindows}
\alias{getMapabilityWindows}
\title{Return coverage object for mapability file based on provided regions}
\usage{
getMapabilityWindows(MapabilityBwFile, regions, ReadLen, WinSz, cores)
}
\arguments{
\item{MapabilityBwFile}{path to the mapability BigWig File}

\item{cores}{number of cores to parallelize using}
}
\value{
mapability coverage object filtered for mapability of "1" regions
}
\description{
Return coverage object for mapability file based on provided regions
}
\examples{
getBAMCoverage("/myproject/myMapability.bw",myGRanges,36)
}

