% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getMappabilityCoverage}
\alias{getMappabilityCoverage}
\title{Return coverage object for mappability file}
\usage{
getMappabilityCoverage(MappabilityBwFile)
}
\arguments{
\item{MappabilityBwFile}{path to the mappability BigWig File}
}
\value{
mappability coverage object filted for mappability of "1" regions
}
\description{
Return coverage object for mappability file
}
\examples{
getBAMCoverage("/myproject/myMappability.bw")
}

