% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getMappabilityGRanges}
\alias{getMappabilityGRanges}
\title{Return normalized mappability based on provided regions}
\usage{
getMappabilityGRanges(MappabilityBwFile, regions, ReadLen, cores)
}
\arguments{
\item{MappabilityBwFile}{path to the mappability BigWig File}

\item{cores}{number of cores to parallelize using}
}
\value{
mappability coverage object filtered for mappability of "1" regions
}
\description{
Return normalized mappability based on provided regions
}
\examples{
getBAMCoverage("/myproject/myMappability.bw",myGRanges,36)
}

