% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BIDCHIPS.R
\name{getMappabilityWindows}
\alias{getMappabilityWindows}
\title{Return coverage object for mappability file based on provided regions}
\usage{
getMappabilityWindows(MappabilityBwFile, regions, ReadLen, WinSz, cores)
}
\arguments{
\item{MappabilityBwFile}{path to the mappability BigWig File}

\item{cores}{number of cores to parallelize using}
}
\value{
mappability coverage object filtered for mappability of "1" regions
}
\description{
Return coverage object for mappability file based on provided regions
}
\examples{
getBAMCoverage("/myproject/myMappability.bw",myGRanges,36)
}

